requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==23.2.1
    # via gradio
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via
    #   instructor
    #   litellm
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   gradio
    #   httpx
    #   openai
    #   starlette
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   litellm
    #   typer
    #   uvicorn
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via instructor
fastapi==0.115.0
    # via gradio
ffmpy==0.4.0
    # via gradio
filelock==3.16.1
    # via huggingface-hub
fonttools==4.54.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==5.0.2
    # via graphs (pyproject.toml)
gradio-client==1.4.0
    # via gradio
graphviz==0.20.3
    # via graphs (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via
    #   gradio
    #   gradio-client
    #   openai
huggingface-hub==0.25.2
    # via
    #   gradio
    #   gradio-client
    #   tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via litellm
instructor==1.5.2
    # via graphs (pyproject.toml)
jinja2==3.1.4
    # via
    #   gradio
    #   litellm
jiter==0.5.0
    # via
    #   instructor
    #   openai
jsonschema==4.23.0
    # via litellm
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
litellm==1.49.1
    # via graphs (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.9.2
    # via
    #   graphs (pyproject.toml)
    #   wordcloud
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.1.2
    # via
    #   contourpy
    #   gradio
    #   matplotlib
    #   pandas
    #   wordcloud
openai==1.51.2
    # via
    #   graphs (pyproject.toml)
    #   instructor
    #   litellm
orjson==3.10.7
    # via gradio
packaging==24.1
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   matplotlib
pandas==2.2.3
    # via gradio
pillow==10.4.0
    # via
    #   gradio
    #   matplotlib
    #   wordcloud
propcache==0.2.0
    # via yarl
pydantic==2.9.2
    # via
    #   graphs (pyproject.toml)
    #   fastapi
    #   gradio
    #   instructor
    #   litellm
    #   openai
pydantic-core==2.23.4
    # via
    #   instructor
    #   pydantic
pydub==0.25.1
    # via gradio
pygments==2.18.0
    # via rich
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via litellm
python-multipart==0.0.12
    # via gradio
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   litellm
    #   tiktoken
rich==13.9.2
    # via
    #   instructor
    #   typer
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
ruff==0.6.9
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.38.6
    # via fastapi
tenacity==8.5.0
    # via instructor
tiktoken==0.8.0
    # via litellm
tokenizers==0.20.1
    # via litellm
tomlkit==0.12.0
    # via gradio
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
typer==0.12.5
    # via
    #   gradio
    #   instructor
typing-extensions==4.12.2
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
uvicorn==0.31.1
    # via gradio
websockets==12.0
    # via gradio-client
wordcloud==1.9.3
    # via graphs (pyproject.toml)
yarl==1.14.0
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata
